+++ src/br/com/example/challengebank/SeniorDeveloper.java
@@ -115,7 +115,7 @@ public class SeniorDeveloper {
 
         System.out.println(menu);
         try {
-            int input = scanner.nextInt();
+            byte input = scanner.nextByte();
             return MenuOption.fromValue(input);
         } catch (InputMismatchException e) {
             System.out.println("Input error.");
